% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_results.R
\name{annotate_results}
\alias{annotate_results}
\title{Annotate DESeq2 results}
\usage{
annotate_results(result, biomart, add, id = 1)
}
\arguments{
\item{result}{a DESeqResults object with Ensembl IDs as row names}

\item{biomart}{annotations from \code{read_biomart} with id column matching
row names in results}

\item{add}{a vector of biomart columns to add to result table, default
gene_name, biotype and description}

\item{id}{position or name of column in biomart that matches rows names in
the DESeq results}
}
\value{
A tibble
}
\description{
Add BioMart annotations to a DESeq2 result table
}
\examples{
\dontrun{
 annotate_results(res, hsa)
}
}
\author{
Chris Stubben
}
