% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.tbl_df.R
\name{as.matrix.tbl_df}
\alias{as.matrix.tbl_df}
\title{Convert tbl_df to a matrix}
\usage{
\method{as.matrix}{tbl_df}(x, ...)
}
\arguments{
\item{x}{a tbl_df with identifiers in column 1 (assigned as rownames) and
numbers in remaining columns}

\item{\dots}{additional options for \code{as.matrix}}
}
\value{
A matrix
}
\description{
Mainly used to convert count tables to matrix for \code{rlog} and other
functions that require a matrix as input.
}
\examples{
\dontrun{
  rlog(as.matrix(fc_tbl))
}
}
\author{
Chris Stubben
}

