% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_matrix.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert tibbles to a matrix}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{a tibble with identifiers in column 1 (assigned as rownames) and
numbers in remaining columns}
}
\value{
A matrix
}
\description{
Mainly used to convert count tables to matrix for Bioconductor functions that
require a matrix as input.
}
\examples{
x <- top_counts(pasilla$results, pasilla$rlog, top=25)
x
as_matrix(x)
}
\author{
Chris Stubben
}
