% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble_gage.R
\name{as_tibble_gage}
\alias{as_tibble_gage}
\title{Convert GAGE output to tibbles}
\usage{
as_tibble_gage(gage_results)
}
\arguments{
\item{gage_results}{Results from \code{gage}}
}
\value{
A list of tibbles
}
\description{
Converts gage ouput (list of matrices) into list of tibbles
and parses matrix rownames into id and name columns.
}
\examples{
\dontrun{
x <- gage(fc, gsets=kegg.sets.mm)
as_tibble_gage(x)
}

}
\author{
Chris Stubben
}

