% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher_all.R
\name{enricher_all}
\alias{enricher_all}
\title{Run enricher from clusterProfiler on all DESeq2 result tables}
\usage{
enricher_all(
  res,
  gsets,
  maxGSSize = 2000,
  expressed_universe = TRUE,
  deseq.padj = 0.05,
  logFC,
  protein_coding = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{res}{A list of DESeq results}

\item{gsets}{A list of gene sets (or two column dataframe with pathway and gene)}

\item{maxGSSize}{Maximum set size, default 2000}

\item{expressed_universe}{Use all expressed genes as universe, default TRUE}

\item{deseq.padj}{Adjusted p-value cutoff for significant genes, default 0.05}

\item{logFC}{absolute fold change cutoff}

\item{protein_coding}{compare protein coding genes only, default TRUE}

\item{quiet}{suppress messages except total significant}

\item{\dots}{Additional options like minGSSize and pvalueCutoff passed to \code{enricher}}
}
\value{
A list of tibbles
}
\description{
Run enricher from clusterProfiler on all DESeq2 result tables
}
\examples{
\dontrun{
library(hciRdata)
enricher_all(res, msig_pathways$KEGG)
}
}
\author{
Chris Stubben
}
