% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{Parse sample from file names}
\usage{
extract_samples(files)
}
\arguments{
\item{files}{file name with path}
}
\value{
A vector
}
\description{
Sample names are parsed from file names without the extension.  If the file name is not unique,
   the parent directory is used.
}
\examples{
# File name or parent directory should be unique
extract_samples(c("align1/1355X1.counts", "align2/1355X2.counts"))
extract_samples(c("align1/1355X1/Log.out", "align2/1355X2/Log.out"))
}
\author{
Chris Stubben
}
