% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Fitler count matrix}
\usage{
filter_counts(count_matrix, min = 10)
}
\arguments{
\item{count_matrix}{a count matrix}

\item{min}{drop feature if all samples have less than or equal to the minimum number of mapped reads, default 10}
}
\value{
A count matrix
}
\description{
Remove rows with zero counts and fewer than a minimum number of mapped reads per sample
}
\examples{
\dontrun{
count_matrix <-  filter_counts(count_matrix,  n=1)
}
}
\author{
Chris Stubben
}

