% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top_counts.R
\name{filter_top_counts}
\alias{filter_top_counts}
\title{Filter top count matrix}
\usage{
filter_top_counts(x, ...)
}
\arguments{
\item{x}{a tibble from \code{link{top_counts}}}

\item{\dots}{additional options passed to \code{subset}}
}
\value{
A count matrix
}
\description{
Remove columns from count matrix returned by \code{link{top_counts}} by subsetting the colData attribute
}
\examples{
\dontrun{
 filter_top_counts( top_counts(res, rld),  type == "paired-end")
}
}
\author{
Chris Stubben
}
