% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_all.R
\name{fisher_all}
\alias{fisher_all}
\title{Run Fisher's test on all DESeq2 result tables}
\usage{
fisher_all(res, gsets, deseq.padj = 0.05, logFC, min_set = 2,
  protein_coding = TRUE)
}
\arguments{
\item{res}{A list of DESeq results}

\item{gsets}{Gene sets with human gene names.}

\item{deseq.padj}{Adjusted p-value cutoff for significant genes, default 0.05}

\item{min_set}{Minimum number of overlaps, default 2}

\item{protein_coding}{compare protein coding genes only, default TRUE}
}
\value{
a tibble with gene sets intersections and pvalue from \code{fisher.test}.
Sets with no intersecting genes are dropped.
}
\description{
Compare significant genes to gene sets using a one-sided Fisher's test.
}
\note{
Assumes gene sets are human genes and will intersect using human_homolog column if present.
Creates a 2 x 2 contingency table with significant and non-significant genes in each gene set and
the p-value is returned from a one-sided \code{fisher.test}
}
\examples{
\dontrun{
library(hciRdata)
fisher_all(res, msig_pathways$KEGG)
}
}
\author{
Chris Stubben
}
