% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_barplot.R
\name{fisher_barplot}
\alias{fisher_barplot}
\title{Gene set enrichment using Fisher's test}
\usage{
fisher_barplot(x, top = 10, ylab = "Significant genes in pathway")
}
\arguments{
\item{res}{A list of DESeq results}

\item{gsets}{Gene sets}

\item{padj}{Adjusted p-value cutoff for significant genes}
}
\value{
a tibble with gene sets intersections and pvalue from \code{fisher.test}
}
\description{
Gene set enrichment using \code{fisher.test}.
}
\examples{
\dontrun{
data(msig)
fisher_all(res, msig$KEGG)
}
}
\author{
Chris Stubben
}
