% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly}
\alias{fly}
\title{Fruitfly gene annotations from Ensembl}
\format{A tibble with 17,559 rows and 10 columns}
\source{
\code{read_biomart("fly")}
}
\usage{
fly
}
\description{
Fruitfly gene annotations from Ensembl
}
\examples{
data(mmu)
mmu
group_by(mmu, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
