% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_msig.R
\name{format_msig}
\alias{format_msig}
\title{Format MsigDB names}
\usage{
format_msig(name)
}
\arguments{
\item{name}{a vector of names}
}
\description{
Formats MsigDB names with all upper case and underlines for spaces
}
\examples{
 format_msig("THIS_IS_AN_MSIGDB_NAME")
}
\author{
Chris Stubben
}
