% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_all.R
\name{gage_all}
\alias{gage_all}
\title{Run GAGE on all DESeq2 result tables}
\usage{
gage_all(res, gsets, fdr = 0.05, ...)
}
\arguments{
\item{res}{DESeq results}

\item{gsets}{gene set}

\item{fdr}{FDR cutoff, default 0.05}

\item{\dots}{Additional options passed to \code{gage}}
}
\value{
A list of tibbles
}
\description{
A wrapper for gage to convert ouput into tibbles
}
\examples{
\dontrun{
data(msig)
x <- gage_all(res, gsets=msig$KEGG)
}

}
\author{
Chris Stubben
}

