% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_coldata.R
\name{gdc_coldata}
\alias{gdc_coldata}
\title{GDC column data}
\usage{
gdc_coldata(gdc)
}
\arguments{
\item{gdc}{a SummarizedExperiment object from \code{GDCprepare} in the \code{TCGAbiolinks} package.}
}
\value{
A tibble
}
\description{
Removes empty and duplicated columns in the \code{colData} slot from \code{GDCprepare}
and returns a tibble.
}
\examples{
\dontrun{
brca <- GDCprepare(query)
meta <- gdc_coldata(brca)
meta
x <- summary_tibble(meta)
x
filter(x, unique == 2)
}
}
\author{
Chris Stubben
}
