% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_rowdata.R
\name{gdc_rowdata}
\alias{gdc_rowdata}
\title{GDC row data}
\usage{
gdc_rowdata(gdc, ensembl)
}
\arguments{
\item{gdc}{a SummarizedExperiment object from \code{GDCprepare} in the \code{TCGAbiolinks} package.}

\item{ensembl}{Ensembl human gene annotations, version 87 is best.}
}
\value{
A tibble
}
\description{
Joins the \code{rowData} slot from \code{GDCprepare} to Ensembl gene annotations
}
\note{
Currently works with harmonized data with Ensembl gene ids.
}
\examples{
\dontrun{
brca <- GDCprepare(query)
data(hsa87)
genes <- gdc_rowdata(brca, hsa87)
genes
}
}
\author{
Chris Stubben
}
