% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_intersect.R
\name{gene_intersect}
\alias{gene_intersect}
\title{Compare genes to sets}
\usage{
gene_intersect(genes, sets, min = 2)
}
\arguments{
\item{genes}{a vector of genes}

\item{sets}{a list of gene sets}

\item{min}{Minimum number of overlaps, default 2}
}
\value{
a tibble with gene set names, size and number of overlapping genes
}
\description{
Intersect genes and a list of gene sets
}
\examples{
gene_intersect(letters[5:8], list(x=letters[3:6], y=letters[6:14], z=letters[10:15]))
}
\author{
Chris Stubben
}
