% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_plotPCA.R
\name{hc_plotPCA}
\alias{hc_plotPCA}
\title{Highchart version of plotPCA in DESeq2}
\usage{
hc_plotPCA(object, intgroup = "condition", tooltip, ntop = 500)
}
\arguments{
\item{object}{a DESeqTransform object}

\item{intgroup}{a character vector of names in colData(x) for grouping, default 'condition'}

\item{tooltip}{a character vector of names in colData(x) for tooltip display,
default displays the object column names'}

\item{ntop}{number of top variable genes to use for principal components}
}
\value{
A highchart
}
\description{
Highchart version of sample PCA plot for transformed data
}
\author{
Chris Stubben
}

