% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsa}
\alias{hsa}
\title{Human gene annotations from Ensembl release 90}
\format{A tibble with 63,967 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcript_count, entrez_gene}
\source{
\code{read_biomart("human")}
}
\usage{
hsa
}
\description{
Human gene annotations from Ensembl release 90
}
\examples{
data(hsa)
hsa
attr(hsa, "downloaded")
group_by(hsa, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
# most ensembl ids have 1 entrez id
table(nchar(gsub("[^,]", "", hsa[["entrez_gene"]]))+1)
}
\keyword{datasets}
