% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_hsa}
\alias{kegg_hsa}
\title{Human pathway gene sets from KEGG}
\format{A named list with 291 pathways with Entrez ID vectors}
\source{
\code{kegg.gsets} in \code{gage} package
}
\usage{
kegg_hsa
}
\description{
Human pathway gene sets from KEGG
}
\note{
Signaling, metabolism and disease pathway indices are saved as attributes
}
\examples{
data(kegg)
kegg_hsa[1:3]
}
\keyword{datasets}

