% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_mmu}
\alias{kegg_mmu}
\title{Mouse pathway gene sets from KEGG}
\format{A named list with 287 pathways with Entrez ID vectors}
\source{
\code{kegg.gsets} in \code{gage} package
}
\usage{
kegg_mmu
}
\description{
Mouse pathway gene sets from KEGG
}
\note{
Signaling, metabolism and disease pathway indices are saved as attributes
}
\examples{
data(kegg_mmu)
length(kegg_mmu)
kegg_mmu[1:3]
names(attributes(kegg_mmu))
length(kegg_mmu[attr(kegg_mmu, "sigmet.idx")] )
}
\keyword{datasets}

