% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu}
\alias{mmu}
\title{Mouse gene annotations from Ensembl}
\format{A tibble with 50,143 rows and 10 columns}
\source{
\code{read_biomart("mouse")}
}
\usage{
mmu
}
\description{
Mouse gene annotations from Ensembl
}
\examples{
data(mmu)
mmu
group_by(mmu, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
}
\keyword{datasets}

