% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig}
\alias{msig}
\title{Canonical pathways in the MSigDB curated gene sets.}
\format{A list with 9 pathways with 1329 gene sets}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\usage{
msig
}
\description{
Nine canonical pathways from the MSigDB C2 curated gene sets.  See \code{\link{read_msigdb}} to load file.
}
\examples{
data(msig)
sapply(msig, length)
msig$KEGG[1:3]
}
\keyword{datasets}
