% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_samples.R
\name{order_samples}
\alias{order_samples}
\title{Order sample names}
\usage{
order_samples(samples)
}
\arguments{
\item{samples}{a vector of sample names}
}
\value{
A vector
}
\description{
Order character vectors by removing non-numeric characters.
}
\examples{
 ids <- c("135X1", "135X21", "135X2", "135X10")
 order(ids)
 order_samples(ids)
}
\author{
Chris Stubben
}

