% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette255.R
\name{palette255}
\alias{palette255}
\title{Expand RColorBrewer palettes}
\usage{
palette255(palette, ramp = TRUE)
}
\arguments{
\item{palette}{RColorBrewer palette name or "RdGn" for Red-Green color scale.}

\item{ramp}{use \code{colorRampPalette}, default returns 255 colors}
}
\value{
A vector of 255 colors, diverging palettes are reversed
}
\description{
Expands palettes into 255 colors  for \code{plot_genes} and \code{plot_volcano}
}
\examples{
x <- palette255("RdGn")
plot(c(1, 256), c(0, 3), type = "n", axes = FALSE, xlab = "", ylab = "")
i <- 1:255
rect(i, 1, i + 1, 2, col = x, border = NA)
}
\author{
Chris Stubben
}
