% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla}
\alias{pasilla}
\title{DESeq objects and results for Pasilla knock-downs}
\format{A list with DESeqDataSet (dds), DESeqTransform (rlog), and tibble (results)}
\source{
\code{pasilla} package and Brooks et al. 2010. Conservation of an RNA regulatory map
between Drosophila and mammals.
}
\usage{
pasilla
}
\description{
DESeq objects and results for Pasilla knock-downs
}
\examples{
data(pasilla)
pasilla
\dontrun{
library("pasilla")
counts   <- read_tsv(system.file("extdata", "pasilla_gene_counts.tsv", package="pasilla") )
samples <- read_csv(system.file("extdata", "pasilla_sample_annotation.csv", package="pasilla" ))
# Need sample data column matching count column names
samples$file <- gsub("fb$", "", samples$file )
## remove 2240 features with 0 reads and 721 with only 1 read
counts  <- filter_counts( counts, sum=TRUE)
dds <- deseq_from_tibble(counts, samples,  design = ~ condition)
rld <- rlog(dds)
fly <- read_biomart("fly")
res <- results_all(dds, fly)
pasilla <- list( dds = dds, rlog = rld, results = res)
}
}
\keyword{datasets}

