% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla_dds}
\alias{pasilla_dds}
\title{DESeqDataSet object with Pasilla knock-downs}
\format{A DESeqDataSet object}
\source{
\code{pasilla} package and Brooks et al. 2010. Conservation of an RNA regulatory map
between Drosophila and mammals.
}
\usage{
pasilla_dds
}
\description{
DESeqDataSet object with Pasilla knock-downs
}
\examples{
data(pasilla_dds)
counts(pasilla_dds)[1:4, ]
\dontrun{
library("pasilla")
count_tbl   <- read_tsv(system.file("extdata", "pasilla_gene_counts.tsv", package="pasilla") )
samples <- read_csv(system.file("extdata", "pasilla_sample_annotation.csv", package="pasilla" ))
# Need sample data column matching count column names
samples$file <- gsub("fb$", "", samples$file )
## remove 2240 features with 0 reads and 721 with only 1 read
count_tbl  <- filter_counts( count_tbl, sum=TRUE)
pasilla_dds <- deseq_from_tibble(count_tbl, samples,  design = ~ condition)
pasilla_dds <- DESeq(pasilla_dds)
}
}
\keyword{datasets}

