% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla_rld}
\alias{pasilla_rld}
\title{Regularized log transform object with Pasilla knock-downs}
\format{A DESeqTransform object}
\source{
\code{pasilla} package and Brooks et al. 2010. Conservation of an RNA regulatory map
between Drosophila and mammals.
}
\usage{
pasilla_rld
}
\description{
Regularized log transform object with Pasilla knock-downs
}
\examples{
data(pasilla_rld)
assay(pasilla_rld)[1:4,]
}
\seealso{
\code{\link{pasilla_dds}} to generate these rlogs
}
\keyword{datasets}

