% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_genes.R
\name{pheatmap_genes}
\alias{pheatmap_genes}
\title{Plot heatmap}
\usage{
pheatmap_genes(x, intgroup, color = "RdYlBu", midpoint0 = TRUE,
  dendsort = TRUE, ...)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{intgroup}{one or more column names for \code{annotation_col} bar}

\item{color}{a divergent color palette name or vector of colors}

\item{midpoint0}{center color scale midpoint at zero}

\item{dendsort}{reorder branches using \code{dendsort} package}

\item{\dots}{additional options passed to \code{pheatmap}}
}
\value{
A pheatmap
}
\description{
Plot pheatmap with annoation bar, reordered branches, and color scale midpoint at zero.
}
\examples{
\dontrun{
 x <- top_counts( res[[1]], rld)
pheatmap_genes(x)
pheatmap_genes(x, "Trt", border=NA )
pheatmap_genes( top_counts(res[[1]], rld , col_label="Name", sort_fc=TRUE)
}
}
\author{
Chris Stubben
}

