% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biotypes.R
\name{plot_biotypes}
\alias{plot_biotypes}
\title{Plot reads assigned to biotypes}
\usage{
plot_biotypes(
  x,
  n = 12,
  group = FALSE,
  stack = "percent",
  sort = TRUE,
  barwidth = NULL,
  ...
)
}
\arguments{
\item{x}{count matrix from featureCounts with `-g gene_biotype`}

\item{n}{number of top features to plot, default 12}

\item{group}{plot all features in 12 groups, default FALSE}

\item{stack, }{plot percent or total reads (percent or normal)}

\item{sort}{}

\item{barwidth}{bar width, default null}

\item{\dots}{additional options passed to \code{hc_size}}
}
\value{
A highchart
}
\description{
Plot reads assigned to biotypes
}
\examples{
\dontrun{
 x <- read_tsv("biotypes.txt")
 plot_biotypes(x)
}
}
\author{
Chris Stubben
}
