% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot counts for a single gene}
\usage{
plot_counts(
  x,
  intgroup,
  n = 25,
  geom = "jitter",
  ylab = "Log2 counts",
  reorder,
  ncol = NULL,
  scales = "fixed",
  ...
)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{intgroup}{column name in \code{attr(x, "colData")} to use for grouping}

\item{n}{total number of genes to plot, default 25}

\item{geom}{geom type, default jitter or point, boxplot and violin}

\item{ylab}{y-axis label}

\item{reorder}{level names to reorder factor levels on x-axis}

\item{ncol}{number of columsn for facet_wrap}

\item{scales}{scales for facet_wrap, default fixed}

\item{\dots}{additional options passed to geom_jitter or other geom}
}
\value{
A ggplot
}
\description{
Plot counts for a single gene
}
\note{
See \code{link{plot_interactions}} to plot many genes.
}
\examples{
 x <- top_counts(pasilla$results, pasilla$rlog, top=12)
 plot_counts(x, "condition")
}
\author{
Chris Stubben
}
