% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot heatmap of sample distances}
\usage{
plot_dist(rld, intgroup, output = "pheatmap", palette = "RdYlBu",
  diagNA = TRUE, font_size = 12, ...)
}
\arguments{
\item{rld}{a matrix or DESeqTransform with assay slot}

\item{intgroup}{one or more column names in colData(rld) for pheatmap \code{annotation_col}}

\item{output}{pheatmap or d3heatmap}

\item{palette}{a color palette name or vector of colors}

\item{diagNA}{set the diagonal to NA, default TRUE}

\item{font_size}{x and yaxis font size for d3heatmap}

\item{\dots}{Additional options like colors passed to \code{d3heatmap} or \code{pheatmap}}
}
\value{
A pheatmap or d3heatmap
}
\description{
Plot sample distances using rlog or other counts.
}
\examples{
data(pasilla_dds)
rld <- rlog(pasilla_dds)
plot_dist(rld, c("condition", "type") )
plot_dist(rld, palette="Blues", diagNA=FALSE)
plot_dist(rld, output = "d3")
}
\author{
Chris Stubben
}

