% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filter.R
\name{plot_filter}
\alias{plot_filter}
\title{Plot count matrix to check filter cutoff}
\usage{
plot_filter(x, n = 20)
}
\arguments{
\item{x}{count matrix, data.frame or tibble}

\item{n}{maximum cutoff, default is to display 0 to 20}
}
\value{
A plot
}
\description{
Plot the total number of features removed from count matrix using different
low count cutoffs for both maximum and total reads
}
\examples{
\dontrun{
 plot_counts(counts)
}
}
\author{
Chris Stubben
}
