% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot heatmap of top genes}
\usage{
plot_genes(x, intgroup, output = "pheatmap", palette = "RdYlBu",
  dendsort = TRUE, midpoint0 = TRUE, border = NA, ...)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{intgroup}{one or more column names for pheatmap \code{annotation_col} bar}

\item{output}{pheatmap or d3heatmap}

\item{palette}{a color palette name or vector of colors}

\item{dendsort}{reorder branches using \code{dendsort} package}

\item{midpoint0}{center color scale midpoint at zero}

\item{border}{pheatmap border color, default none}

\item{\dots}{additional options passed to \code{pheatmap}}
}
\value{
A pheatmap
}
\description{
Plot heatmap of top genes with annoation bars, reordered branches, and color scale midpoint at zero.
}
\examples{
\dontrun{
 x <- top_counts( res, rld)
plot_genes(x)
plot_genes(x, "Trt", palette="RdBu" )
plot_genes( top_counts(res, rld, sort_fc=TRUE) )
}
}
\author{
Chris Stubben
}

