% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot GSEA output}
\usage{
plot_gsea(x, trim = 70, n_sets, nes = TRUE, ...)
}
\arguments{
\item{x}{list with two or more \code{link{read_gsea}} tables}

\item{trim}{trim long set names, default more than 70 characters}

\item{n_sets}{display contrasts sharing n or more sets for n > 1.  If n = 1,
then only plot unique sets.  If missing, then plots all sets, default.}

\item{nes}{plot NES (or ES if FALSE)}

\item{\dots}{other options passed to \code{pheatmap}}
}
\description{
Plot heatmap with enrichment scores by contrast
}
\examples{
\dontrun{
data(msig)
x <- list(X1 = read_gsea("gsea1.txt"), X2 = read_gsea("gsea2.txt"))
plot_gsea(x)
}
}
\author{
Chris Stubben
}
