% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Highchart version of plotPCA in DESeq2}
\usage{
plot_pca(object, intgroup = "condition", tooltip, ntop = 500, pc = c(1,
  2), ...)
}
\arguments{
\item{object}{a DESeqTransform object}

\item{intgroup}{a character vector of names in colData(x) for grouping, default 'condition'}

\item{tooltip}{a character vector of names in colData(x) for tooltip display,
default displays the object column names}

\item{ntop}{number of top variable genes to use for principal components}

\item{pc}{a vector of components to plot, default 1st and 2nd}

\item{\dots}{additional options passed to \code{hc_chart}}
}
\value{
A highchart
}
\description{
Highchart version of sample PCA plot for transformed data
}
\examples{
dds <- makeExampleDESeqDataSet(betaSD=1)
rld <- rlog(dds)
plotPCA(rld)
plot_pca(rld)
colData(rld)$sample <- 1:12
colData(rld)$gender <- sample(c("M", "F"), 12, TRUE)
plot_pca(rld, tooltip=c("sample", "gender"))
}
\author{
Chris Stubben
}

