% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_dist.R
\name{plot_sample_dist}
\alias{plot_sample_dist}
\title{Plot heatmap of sample distances}
\usage{
plot_sample_dist(rld, intgroup, output = "pheatmap", palette = "RdYlBu",
  revclrs = FALSE, diagNA = TRUE, font_size = 12, ...)
}
\arguments{
\item{rld}{a matrix or DESeqTransform with assay slot}

\item{intgroup}{a character vector of names in colData(rld) for pheatmap \code{annotation_col}}

\item{output}{pheatmap or d3heatmap}

\item{palette}{RColorBrewer palette name}

\item{revclrs}{reverse color palette}

\item{diagNA}{set the diagonal to NA, default TRUE}

\item{font_size}{x and yaxis font size for d3heatmap}

\item{\dots}{Additional options like colors passed to \code{d3heatmap} or \code{pheatmap}}
}
\value{
A pheatmap or d3heatmap
}
\description{
Plot sample distances using rlog or other counts.
}
\examples{
data(pasilla_dds)
rld <- rlog(pasilla_dds)
plot_sample_dist(rld, c("condition", "type") )
plot_sample_dist(rld, palette="Blues", revclrs=TRUE, diagNA=FALSE)
plot_sample_dist(rld, output = "d3")
}
\author{
Chris Stubben
}

