% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_genes.R
\name{print_genes}
\alias{print_genes}
\title{Print the scaled values from plot_genes and optionally add branches from cutree}
\usage{
print_genes(x, cut)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{cut}{number of branches to cut}
}
\value{
tibble with id, branch and rlog values used by \code{\link{plot_genes}}
}
\description{
Print the scaled values from plot_genes and optionally add branches from cutree
}
\examples{
x <- top_counts(pasilla$results, pasilla$rlog)
plot_genes(x, c("condition", "type"), scale="row", annotation_names_col=FALSE)
print_genes(x, cut=5)
}
\author{
Chris Stubben
}
