% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomart.R
\name{read_biomart}
\alias{read_biomart}
\title{Read biomart annotations}
\usage{
read_biomart(dataset = "human", fragments = TRUE)
}
\arguments{
\item{dataset}{ensembl dataset name from \code{listEnsembl}}

\item{fragments}{drop fragments like GL456211.1, CHR_MG132_PATCH (any chromosome
name with 3 or more characters), default TRUE}
}
\value{
A tibble with id, biotype, description, chromosome, start, end, strand
 gene name and transcript count
}
\description{
Read Ensembl annotations using \code{biomaRt} library
}
\examples{
\dontrun{
   mm <- read_biomart("mouse")
}
}
\author{
Chris Stubben
}

