% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deseq.R
\name{read_deseq}
\alias{read_deseq}
\title{Read a DESeq Excel file from write_deseq}
\usage{
read_deseq(file, object = "results")
}
\arguments{
\item{file}{write_deseq output file}

\item{object}{type of object to create, either results, rlog or log2_norm}
}
\value{
a list of tibbles or DESeqTransform object
}
\description{
Read a DESeq Excel file from write_deseq
}
\examples{
\dontrun{
 res <- read_deseq("DESeq.xlsx")
 rld <- read_deseq("DESeq.xlsx", "rlog")
}
}
\author{
Chris Stubben
}
