% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_featureCounts.R
\name{read_featureCounts}
\alias{read_featureCounts}
\title{Read featureCounts output files}
\usage{
read_featureCounts(path = ".", pattern, reshape = TRUE, stats = FALSE)
}
\arguments{
\item{path}{the path to featureCounts output files, the default
corresponds to the working directory.}

\item{pattern}{regular expression for file name matching, deafult .counts and .summary}

\item{reshape}{reshape into wide format with samples in rows (count matrix)}

\item{stats}{read stats tables, default counts}
}
\value{
A tibble
}
\description{
Reads featureCount count or alignment summary files and optionally reshape into
 wide format.
}
\examples{
\dontrun{
 ft_counts <- read_featureCounts(".")
 fc <- read_featureCounts(".", stats=TRUE)
 fc
 fc <- read_featureCounts(".", stats=TRUE, reshape=FALSE)
 fc
 filter(fc, count!=0) \%>\%
   hchart("bar", x=sample ,  y=count, group=status) \%>\%
     hc_plotOptions(bar = list(stacking = "normal"))
}
}
\author{
Chris Stubben
}

