% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gsea.R
\name{read_gsea}
\alias{read_gsea}
\alias{read_gsea_file}
\title{Read GSEA enrichment files}
\usage{
read_gsea(gsea_dir, FDR = 0.1)

read_gsea_file(gsea_file)
}
\arguments{
\item{gsea_dir}{Path to GSEA output files}

\item{FDR}{FDR cutoff, default 0.1}

\item{gsea_file}{GSEA output file}
}
\value{
A tibble
}
\description{
Reads both positive and negative enrichment results from Broad's GSEA into a single table
}
\section{Functions}{
\itemize{
\item \code{read_gsea_file()}: Read a single GSEA output file

}}
\examples{
\dontrun{
x <- read_gsea("~/gsea_home/output/dec15/my_analysis.GseaPreranked.1513361740375")
}
}
\author{
Chris Stubben
}
