% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idxstats.R
\name{read_idxstats}
\alias{read_idxstats}
\title{Read samtools idxstats output}
\usage{
read_idxstats(path = ".", pattern, reshape = FALSE, fragment = FALSE)
}
\arguments{
\item{path}{the path to idxstats output files, the default
corresponds to the working directory.}

\item{pattern}{regular expression for file name matching, default .idxstats}

\item{reshape}{reshape mapped reads into wide format with samples in rows}

\item{fragment}{keep sequence names like GL000008.2 with 8 or more characters}
}
\value{
A tibble
}
\description{
Read mapped reads by sequence in samtools idxstats output
}
\examples{
\dontrun{
   read_idxstats( pattern = "idxstats.txt$")
}
}
\author{
Chris Stubben
}

