% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ipa.R
\name{read_ipa}
\alias{read_ipa}
\title{Read IPA Expression Analysis output file}
\usage{
read_ipa(file, excel = FALSE, mylists = FALSE)
}
\arguments{
\item{file}{IPA output file (created using Export All on the main Summary tab)}

\item{excel}{write to Excel file}

\item{mylists}{include optional My Pathways and My Lists}
}
\value{
a list of tibbles or Excel file
}
\description{
Read an IPA Expression Analysis output text file and optionally
write to an  Excel file with each table in a separate tab.
}
\examples{
 \dontrun{
 x <- read_ipa("RPR2_vs_RPM.txt")
 names(x)
 read_ipa("RPR2_vs_RPM.txt",  excel=TRUE)
 }
}
\author{
Chris Stubben
}

