% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msigdb.R
\name{read_msigdb}
\alias{read_msigdb}
\title{Read MSigDB GMT file into a nested list of databases and genes sets}
\usage{
read_msigdb(gmt)
}
\arguments{
\item{gmt}{a GMT file from MSigDB}
}
\value{
A nested list of databases and genes sets (if only one database in GMT file,
then a list of gene sets only)
}
\description{
Reads MSigDB GMT files from \url{http://software.broadinstitute.org/gsea/msigdb/collections.jsp}
}
\examples{
 gmt <- system.file("extdata", "c2.cp.v6.0.symbols.gmt", package = "hciR2")
 msig <- read_msigdb(gmt)
 # a list with 9 pathways dbs
 sapply(msig, length)
 msig$KEGG[1:3]
}
\author{
Chris Stubben
}

