% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_files.R
\name{read_sample_files}
\alias{read_sample_files}
\title{Read and combine common output files}
\usage{
read_sample_files(path = ".", pattern = "\\\\.counts$", delim = "\\t",
  ...)
}
\arguments{
\item{path}{the path to output files, the default corresponds to the working directory.}

\item{pattern}{regular expression for file name matching}

\item{delim}{separator in output files, default table}

\item{\dots}{additional options such as col_names passed to \code{read_delim}.}
}
\value{
A list with coverage and stats data.frames
}
\description{
Read output files and parse the file name to add sample IDs in the first column
}
\note{
Sample names are parsed from file names without extensions.  If the file name is not unique,
   the parent directory is used.   Requires tibble > version 1.2 to avoid error in add_column
}
\examples{
\dontrun{
   #FeatureCounts summary (second column name with *.bam is always unique, so skip and assign)
   fc <- read_sample_files(".summary$", skip=1, col_names=c("status", "count"))
 filter(fc, count!=0) \%>\%
   hchart("bar", x=sample, y=count, group=status) \%>\%
    hc_plotOptions(bar = list(stacking = "normal"))
}
}
\author{
Chris Stubben
}

