% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_STAR.R
\name{read_STAR}
\alias{read_STAR}
\title{Read STAR log files}
\usage{
read_STAR(path = ".", pattern, reshape = FALSE)
}
\arguments{
\item{path}{the path to STAR log files, the default corresponds to the working directory.}

\item{pattern}{regular expression for file name matching, default .final.out}

\item{reshape}{reshape percent mapping into wide format with samples in rows}
}
\value{
A tibble
}
\description{
Read STAR Log.final.out files and optionally reshape into wide format.
}
\note{
Reading output files requires a unique sample identifier in either the file name or parent directory
}
\examples{
\dontrun{
 x <- read_STAR( pattern=".star.out$")
  reads <- c("Uniquely mapped",  "Mapped to multiple loci",
               "Mapped to too many loci", "Unmapped reads")
  y <- filter(x, stat \%in\% reads) \%>\% mutate( stat= factor(stat, levels=reads ))
         hchart(y, "bar", x=sample ,  y=value, group=stat ) \%>\%
            hc_plotOptions(bar = list(stacking = "normal")) \%>\%
             hc_colors( c('#437bb1', '#7cb5ec', '#f7a35c', '#b1084c') ) \%>\%
               hc_yAxis(reversedStacks = FALSE)
 read_STAR( pattern=".star.out$", reshape=TRUE)
}
}
\author{
Chris Stubben
}
