% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results2.R
\name{results2}
\alias{results2}
\title{Extract and annotate a single DESeq results table}
\usage{
results2(dds, biomart, exp, ref, trt = "trt", alpha = 0.05,
  lfcShrink = TRUE)
}
\arguments{
\item{dds}{a DESeqDataSet}

\item{biomart}{annotations from \code{read_biomart} with column 1 matching row names in results}

\item{exp}{name of experimental trt group in the numerator}

\item{ref}{name of reference trt group in the denominator}

\item{trt}{treatment group in the design formula}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{lfcShrink}{shrink fold changes using \code{lfcShrink} for DESeq2 version >= 1.16}
}
\value{
A tibble
}
\description{
Extract and annotate a single contrast.  See \code{\link{results_all}} to annotate all contrasts.
}
\examples{
\dontrun{
library(hciRdata)
res <- results2(pasilla$dds, fly98, "treated", "untreated", trt="condition")
}
}
\author{
Chris Stubben
}
