% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_apeglm.R
\name{results_apeglm}
\alias{results_apeglm}
\title{Run all results from a DESeq analysis using apeglm shrinkage}
\usage{
results_apeglm(
  object,
  biomart,
  alpha = 0.05,
  subset,
  add_columns,
  trt,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{biomart}{annotations from \code{read_biomart} with column 1 matching row names in results}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{subset}{sorted index to subset all pairwise comparisons, see check_contrasts for possible contrasts}

\item{add_columns}{a vector of biomart columns to add to result table, default
gene_name, biotype, chromosome, description and human_homolog if present}

\item{trt}{Compare groups within trt group, default is first term in the design formula}

\item{simplify}{return a tibble if only 1 contrast present}

\item{\dots}{additional options passed to \code{results}}
}
\value{
A list of tibbles for each contrast
}
\description{
Run all results from a DESeq analysis using apeglm shrinkage
}
\examples{
\dontrun{
library(hciRdata)
res <- results_apeglm(pasilla$dds, fly98)
res
}
}
\author{
Chris Stubben
}
