% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlog_dds.R
\name{rlog_dds}
\alias{rlog_dds}
\title{Regularized log transformation}
\usage{
rlog_dds(dds, ...)
}
\arguments{
\item{a}{DESeqDataSet}

\item{\dots}{additional options passed to \code{rlog}}
}
\value{
A DESeqTransform object
}
\description{
This is a wrapper for \code{rlog} and avoids loading the DESeq2 package.
}
\examples{
\dontrun{
  rlog_dds( dds)
}
}
\author{
Chris Stubben
}

