% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_names.R
\name{sample_names}
\alias{sample_names}
\title{Parse sample from file names}
\usage{
sample_names(files)
}
\arguments{
\item{files}{file name with path}
}
\value{
A vector
}
\description{
Sample names are parsed from file names without the extension.  If the file name is not unique,
   the parent directory is used.
}
\examples{
 # File name or parent directory should be unique
 sample_names(c("align1/1355X1.counts", "align2/1355X2.counts"))
 sample_names(c("align1/1355X1/Log.out", "align2/1355X2/Log.out"))
}
\author{
Chris Stubben
}

