% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sort.R
\name{sample_sort}
\alias{sample_sort}
\title{Sort sample names}
\usage{
sample_sort(samples)
}
\arguments{
\item{samples}{a vector of sample names}
}
\value{
A vector
}
\description{
Sort character vectors by removing non-numeric characters.  This works for very simple cases,
 try \code{mixedsort} in \code{ggtools} for more complicated examples
}
\examples{
 samples <- c("135X1", "135X2", "135X11")
 sort(samples)
 sample_sort(samples)
}
\author{
Chris Stubben
}

