% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_counts.R
\name{sort_counts}
\alias{sort_counts}
\title{Sort the columns of a count matrix by rows in sample data}
\usage{
sort_counts(count_matrix, sample_data, id = 1)
}
\arguments{
\item{count_matrix}{a count matrix}

\item{sample_data}{a sample table}

\item{id}{position of column in sample data that matches column names in the
count matrix, default is the first column}
}
\value{
A re-ordered count matrix
}
\description{
Creating a DESeqDataSet from a count matrix requires columns sorted by the rows in the sample dataset.
 This reorders the columns in the count matrix to match sample data.
}
\examples{
\dontrun{
count_matrix <-  sort_counts(count_matrix, sample_data)
}
}
\author{
Chris Stubben
}

