% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_samples.R
\name{sort_samples}
\alias{order_samples}
\alias{sort_samples}
\title{Sort sample names}
\usage{
sort_samples(samples)

order_samples(samples)
}
\arguments{
\item{samples}{a vector of sample names}
}
\value{
A vector
}
\description{
Sort character vectors by removing non-numeric characters.  This works for very simple names,
 try \code{mixedsort} in \code{ggtools} for more complicated examples
}
\section{Functions}{
\itemize{
\item \code{order_samples}: Order sample names
}}
\examples{
 ids <- c("135X1", "135X21", "135X2", "135X10")
 sort(ids)
 sort_samples(ids)
 order_samples(ids)
}
\author{
Chris Stubben
}

