% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_DESeq.R
\name{summary_DESeq}
\alias{summary_DESeq}
\title{Summarize DESeq2 results}
\usage{
summary_DESeq(object)
}
\arguments{
\item{object}{a DESeqResults object}
}
\value{
A data.frame
}
\description{
Summarize DESeq2 results into a \code{data.frame}
}
\examples{
\dontrun{
summary_DESeq(res1)
}
}
\author{
Chris Stubben
}

