% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tibble.R
\name{summary_tibble}
\alias{summary_tibble}
\alias{top3}
\alias{drop_empty_columns}
\title{Summarize tibble columns}
\usage{
summary_tibble(x)

top3(y)

drop_empty_columns(x)
}
\arguments{
\item{x}{a tibble}

\item{y}{a vector}
}
\value{
A tibble with column names, types, number of unique values and NAs,
minimum, maximum and top three values.
}
\description{
Summarize tibbles with many columns like TCGA metadata.  Summaries for large tibbles
with many rows will be slow and not recommended.
}
\section{Functions}{
\itemize{
\item \code{top3}: Top three values

\item \code{drop_empty_columns}: Drop empty columns
}}

\examples{
 summary_tibble(mtcars)
 #drop_empty_columns( tibble( a=1:5, b=NA, c="x", d="") )
}
\author{
Chris Stubben
}
