% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{Get top genes}
\usage{
top_genes(res, top = 40, padjust = 0.05, basemean, log2FC, sort_fc = FALSE)
}
\arguments{
\item{res}{an annotated DESeq2 results file}

\item{top}{Number of top genes to display in matrix}

\item{padjust}{Adjusted p-value cutoff}

\item{basemean}{basemean cutoff}

\item{log2FC}{absolute value log2 fold change cutoff}

\item{sort_fc}{Get the top n/2 up and down-regulated genes sorted by adjusted p-value}
}
\value{
A tibble of top genes
}
\description{
Return top genes in DESeq results
}
\examples{
top_genes(pasilla$results) \%>\% dplyr::select(1:7,11)
}
\author{
Chris Stubben
}
