% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_deseq.R
\name{write_deseq}
\alias{write_deseq}
\title{Write DESeq results to an Excel file}
\usage{
write_deseq(dds, result_all, rld, biomart, file = "DESeq.xlsx", ...)
}
\arguments{
\item{dds}{a DESeqDataset objects object with Ensembl IDs as row names}

\item{result_all}{a list from \code{results_all}}

\item{rld}{a DESeqTransform obect with rlog values}

\item{biomart}{annotations from \code{read_biomart}}

\item{file}{file name}

\item{\dots}{additional options passed to \code{annotate_results}}
}
\value{
A tibble
}
\description{
Write DESeq result files, raw counts, rlog values, normalized counts.
}
\examples{
\dontrun{
 write_deseq(dds, res, rld, hsa)
}
}
\author{
Chris Stubben
}

