% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gsea_rnk.R
\name{write_gsea_rnk}
\alias{write_gsea_rnk}
\title{Write DESeq results to GSEA rank file}
\usage{
write_gsea_rnk(res, txt = TRUE)
}
\arguments{
\item{res}{a list of DESeq results}

\item{txt}{add .txt ending to contrast.rnk for GNomEx}
}
\value{
Tab-delimited file with gene name and log2 fold change
}
\description{
Writes gene name (or human homolog) and log2 fold change sorted in
 descending order to tab-delimited file
}
\examples{
\dontrun{
  write_gsea_rnk(res)
}
}
\author{
Chris Stubben
}
