% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ipa.R
\name{write_ipa}
\alias{write_ipa}
\title{Write DESeq results to text files for bulk uploading into IPA}
\usage{
write_ipa(result_all)
}
\arguments{
\item{result_all}{a list from \code{results_all}}
}
\value{
Text files
}
\description{
Write DESeq results to text files for bulk uploading into IPA
}
\examples{
\dontrun{
 write_ipa(res)
}
}
\author{
Chris Stubben
}
