% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_contrasts.R
\name{check_contrasts}
\alias{check_contrasts}
\title{Check contrasts}
\usage{
check_contrasts(trt, vs = "all", vs2 = TRUE, control_first = TRUE)
}
\arguments{
\item{trt}{a vector}

\item{vs}{either compare all vs. all (default) or a all vs. a specific treatment, or see note.}

\item{vs2}{position of specific treatment in contrast vector, set FALSE for specific treatment vs all}
}
\value{
A vector of contrasts
}
\description{
Check all possible contrasts used by \code{\link{results_all}}
}
\note{
If you combine factors of interest into a single group following section 3.3 in the DESeq2 vignette,
 you can set vs = "combined" to limit the comparisons to those within the first group.
}
\examples{
 # check_contrasts( samples$trt)
 trt <- factor(c( "control", "heart", "lung"))
 check_contrasts(trt)
 trt <- factor(trt, levels = c(  "heart", "lung", "control"))
 check_contrasts(trt, control_first =FALSE)
 data.frame( vs=check_contrasts(trt), control_first =FALSE)
 check_contrasts(trt, vs = "control")
 check_contrasts(trt, vs = "control", vs2=FALSE)
 check_contrasts(factor(trt, levels = trt)  )
 # combine 2 trt groups  (66 possible contrasts, 12 within 1st group).
 trt <- paste( 1:4, rep(c("A", "B", "C"), each=4))
 check_contrasts( trt, vs = "combined")
}
\author{
Chris Stubben
}
