% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_from_tibble.R
\name{deseq_from_tibble}
\alias{deseq_from_tibble}
\title{Run DESeq from tibbles}
\usage{
deseq_from_tibble(counts, samples, run_deseq = TRUE, minReplicates = 7, ...)
}
\arguments{
\item{counts}{a count table with feature ids in row 1}

\item{samples}{a sample table with names matching count table column names in row 1}

\item{run_deseq}{Run \code{\link{DESeq}}, default TRUE}

\item{minReplicates}{minimum number of replicates required in order to use
replaceOutliers on a sample, default 7.}

\item{\dots}{additional options like the design formula are passed to \code{DESeqDataSetFromMatrix}}
}
\value{
A DESeqDataSet object
}
\description{
Creates DESeqDataSet object using count and sample tibbles as input and optionally runs DESeq
}
\note{
This function first runs \code{\link{sort_counts}} to check and
reorder count columns by the first column in samples, and then \code{DESeqDataSetFromMatrix}
and optionally \code{DESeq}.  To match DESeq2 versions < 1.16, set run_deseq = FALSE and then
run \code{DESeq(dds, betaPrior =TRUE)}'
}
\examples{
\dontrun{
   deseq_from_tibble(counts, samples, design = ~ trt)
}
}
\author{
Chris Stubben
}
