% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_all.R
\name{fgsea_all}
\alias{fgsea_all}
\title{Run fgsea on all DESeq2 result tables}
\usage{
fgsea_all(res, gsets, FDR = 0.1, nperm = 10000, ...)
}
\arguments{
\item{res}{A list of DESeq results}

\item{gsets}{Gene sets}

\item{FDR}{FDR cutoff, default 0.1}

\item{nperm}{Number of permutations, default 10000}

\item{\dots}{Additional options passed to \code{fgsea}}
}
\value{
A list of tibbles
}
\description{
A wrapper for fgsea to convert ouput into tibbles
}
\examples{
\dontrun{
library(hciRdata)
fc <- write_gsea_rnk(res, write=FALSE)
x <- fgsea_all(fc, gsets=msig_pathways$KEGG)
}

}
\author{
Chris Stubben
}
