% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_counts.R
\name{filter_counts}
\alias{filter_counts}
\title{Filter count matrix}
\usage{
filter_counts(count_tbl, n = 1, sum = FALSE)
}
\arguments{
\item{count_tbl}{a count matrix, data.frame or tibble}

\item{n}{low count cutoff, default 1}

\item{sum}{Use total read counts to apply the cutoff.  The default is the maximum reads.}
}
\value{
A count matrix
}
\description{
Remove features with zero counts and less than the low count cutoff.  The
 default method is a maximum-based filter unless sum=TRUE
}
\examples{
c1 <- matrix(c(0,0,0,2,12,0,0,1,1,0,0,1,1,1,0), ncol=3)
c1
filter_counts(c1)
filter_counts(c1, sum=TRUE)
}
\author{
Chris Stubben
}
