% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_all.R
\name{fisher_all}
\alias{fisher_all}
\title{Run Fisher's test on all DESeq2 result tables}
\usage{
fisher_all(res, gsets, deseq.padj = 0.05, min = 2)
}
\arguments{
\item{res}{A list of DESeq results}

\item{gsets}{Gene sets}

\item{deseq.padj}{Adjusted p-value cutoff for significant genes, default 0.05}

\item{min}{Minimum number of overlaps, default 2}
}
\value{
a tibble with gene sets intersections and pvalue from \code{fisher.test}.
Sets with no intersecting genes are dropped.
}
\description{
Compare significant genes to gene sets using a one-sided Fisher's test.
}
\note{
This functions removes non-coding genes in results and creates a 2 x 2
contingency table with significant and non-significant genes in each gene set.
A one-sided \code{fisher.test} returns p-values for all sets with at least
the minimum number of overlapping genes.
}
\examples{
\dontrun{
data(msig)
fisher_all(res, msig$KEGG)
}
}
\author{
Chris Stubben
}
