% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_barplot.R
\name{fisher_barplot}
\alias{fisher_barplot}
\title{Gene set enrichment using Fisher's test}
\usage{
fisher_barplot(x, top = 10, ylab = "Significant genes in pathway")
}
\arguments{
\item{x}{output from \code{\link{fisher_all}}}

\item{top}{number of top genes}

\item{ylab}{y-axis label}
}
\value{
a ggplot
}
\description{
Gene set enrichment using \code{fisher.test}.
}
\examples{
\dontrun{
data(msig)
fisher_all(res, msig$KEGG)
}
}
\author{
Chris Stubben
}
