% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdc_coldata.R
\name{gdc_coldata}
\alias{gdc_coldata}
\title{GDC column summary}
\usage{
gdc_coldata(gdc)
}
\arguments{
\item{gdc}{a SummarizedExperiment object from \code{GDCprepare}}
}
\value{
A tibble
}
\description{
Summarize the \code{colData} slot from \code{GDCprepare} in \code{TCGAbiolinks}
}
\examples{
\dontrun{
brca <- GDCprepare(query)
meta <- gdc_meta(brca)
meta
x <- summary_tibble(meta)
x
filter(x, unique==2)
}
}
\author{
Chris Stubben
}
