% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_hsa}
\alias{kegg_hsa}
\title{Human pathway gene sets from KEGG}
\format{A tibble with 25,059 rows and 291 pathways}
\source{
\code{kegg.gsets} in \code{gage} package
}
\usage{
kegg_hsa
}
\description{
Human pathway gene sets from KEGG
}
\note{
Signaling, metabolism and disease pathway indices are saved as attributes
}
\examples{
data(kegg_hsa)
kegg_hsa
names(attributes(kegg_hsa))
filter(kegg_hsa, id \%in\% attr(kegg_hsa, "dise.idx")) \%>\%
  group_by(pathway) \%>\% summarize(n = n()) \%>\% arrange(desc(n))
\dontrun{
library(gage)
x <- kegg.gsets(species = "hsa")
n <- sapply(x$kg.sets, length)
y <- names(x$kg.sets)
kegg_hsa <- tibble( id = rep(1:length(y), n),  entry = rep(substr(y, 1,8), n),
  pathway = rep(substring(y, 10),n) , entrez = unlist(x$kg.sets) )
for(i in 2:5) attr(kegg_hsa, names(x1)[i]) <- x1[[i]]
}
}
\keyword{datasets}

