% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_hsa}
\alias{kegg_hsa}
\title{Human pathway gene sets from KEGG}
\format{A tibble with 25,059 rows and 291 pathways}
\source{
\code{kegg.gsets} in \code{gage} package
}
\usage{
kegg_hsa
}
\description{
Human pathway gene sets from KEGG
}
\note{
Signaling, metabolism and disease pathway indices are saved as attributes
}
\examples{
data(kegg_hsa)
kegg_hsa
names(attributes(kegg_hsa))
filter(kegg_hsa, id \%in\% attr(kegg_hsa, "dise.idx")) \%>\%
  group_by(pathway) \%>\% summarize(n = n()) \%>\% arrange(desc(n))
}
\keyword{datasets}
