% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_mmu}
\alias{kegg_mmu}
\title{Mouse pathway gene sets from KEGG}
\format{A tibble with 26,255 rows and 288 pathways}
\source{
\code{kegg.gsets} in \code{gage} package
}
\usage{
kegg_mmu
}
\description{
Mouse pathway gene sets from KEGG
}
\note{
Signaling, metabolism and disease pathway indices are saved as attributes
}
\examples{
data(kegg_mmu)
kegg_mmu
names(attributes(kegg_mmu))
filter(kegg_mmu, id \%in\% attr(kegg_mmu, "dise.idx")) \%>\%
  group_by(pathway) \%>\% summarize(n = n()) \%>\% arrange(desc(n))
}
\keyword{datasets}

