% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgi}
\alias{mgi}
\title{Human and Mouse Homologs}
\format{A tibble with 18,463 rows and 6 variables:
\describe{
  \item{Human}{ Human Marker Symbol }
  \item{EntrezGene}{ Human Entrez Gene ID }
  \item{HomoloGene }{ HomoloGene ID }
  \item{Mouse}{ Mouse Marker Symbol }
  \item{MGI}{ MGI Marker Accession ID }
  \item{PhenotypeId}{ High-level Mammalian Phenotype ID  }
}}
\source{
\url{http://www.informatics.jax.org/downloads/reports/HMD_HumanPhenotype.rpt}
}
\usage{
mgi
}
\description{
Human and Mouse Homology with phenotype annotations from MGI
}
\examples{
data(mgi)
mgi
table( mgi$Human == toupper(mgi$Mouse) )  # 85.7\% are the same
}
\keyword{datasets}
