% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu}
\alias{mmu}
\title{Mouse gene annotations from Ensembl release 90}
\format{A tibble with 53,110 rows and 11 columns}
\source{
\code{read_biomart("mouse")}. Human homologs are added from Ensembl
using the maximum percent identity.  See
 the \code{mmu.R} file in the \code{data-raw} directory for details.
}
\usage{
mmu
}
\description{
Mouse gene annotations from Ensembl release 90
}
\examples{
data(mmu)
mmu
group_by(mmu, biotype) \%>\%
 summarize(n=n(), human_homologs = sum(!is.na(human_homolog))) \%>\%
  arrange(desc(n))
}
\keyword{datasets}
