% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu}
\alias{mmu}
\title{Mouse gene annotations from Ensembl}
\format{A tibble with 50143 rows and 10 columns: id, gene_name, biotype, chromosome,
start, end, strand, description, transcript_count, entrez_gene}
\source{
\code{read_biomart("mouse")}
}
\usage{
mmu
}
\description{
Mouse gene annotations from Ensembl
}
\examples{
data(mmu)
mmu
attr(mmu, "downloaded")
group_by(mmu, biotype) \%>\% summarize(n=n()) \%>\% arrange(desc(n))
# most ensembl ids have 1 entrez id
table(nchar(gsub("[^,]", "", mmu[["entrez_gene"]]))+1)
}
\keyword{datasets}

