% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasilla}
\alias{pasilla}
\title{DESeq objects and results for Pasilla knock-downs}
\format{
A list with DESeqDataSet (dds), DESeqTransform (rlog), and tibble (results)
}
\source{
\code{pasilla} package and Brooks et al. 2010. Conservation of an RNA regulatory map
between Drosophila and mammals.
}
\usage{
pasilla
}
\description{
DESeq objects and results for Pasilla knock-downs
}
\examples{
data(pasilla)
pasilla
}
\keyword{datasets}
