% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot counts for a single gene}
\usage{
plot_counts(rld, gene, intgroups, ylab = "count", title)
}
\arguments{
\item{rld}{rlog or other counts in DESeqTransform object}

\item{gene}{gene id, should match rownames in rld}

\item{intgroups}{one or two names in colData(rld) to use for grouping}

\item{ylab}{y-axis label}

\item{title}{ggplot title, use NULL for no title}
}
\value{
A ggplot
}
\description{
Plot counts for a single gene
}
\note{
See \code{link{plot_interactions}} to plot many genes.
}
\examples{
\dontrun{
 DESeq2::plotCounts(dds, "ENSG00000075624",  "time")
 plot_counts(rld, "ENSG00000075624",  "time")
 # interaction plot
 plot_counts(rld, "ENSG00000075624",  c("time", "trt"), title = "ACTB")
}
}
\author{
Chris Stubben
}
