% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot heatmap of sample distances}
\usage{
plot_dist(rld, intgroup, col_name, output = "pheatmap",
  palette = "RdYlBu", reverse_pal = FALSE, diagNA = TRUE,
  border = NA, fontsize = 10, ...)
}
\arguments{
\item{rld}{a matrix or DESeqTransform with assay slot}

\item{intgroup}{one or more column names in colData(rld) for pheatmap \code{annotation_col}}

\item{col_name}{Column name in `colData(rld)` for labels, default is sample IDs in colnames(rld)}

\item{output}{pheatmap or d3heatmap}

\item{palette}{a color palette name or vector of colors}

\item{reverse_pal}{reverse the color palette}

\item{diagNA}{set the diagonal to NA, default TRUE}

\item{border}{pheatmap border color, default none}

\item{fontsize}{x and yaxis font size}

\item{\dots}{Additional options like colors passed to \code{d3heatmap} or \code{pheatmap}}
}
\value{
A pheatmap or d3heatmap
}
\description{
Plot sample distances using rlog or other counts.
}
\examples{
data(pasilla)
plot_dist(pasilla$rlog, c("condition", "type") )
plot_dist(pasilla$rlog, palette="Blues", diagNA=FALSE)
plot_dist(pasilla$rlog, output = "d3")
}
\author{
Chris Stubben
}
