% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gage.R
\name{plot_gage}
\alias{plot_gage}
\title{Plot gage_all output}
\usage{
plot_gage(x, ...)
}
\arguments{
\item{x}{list from gage_all}

\item{\dots}{other options passed to \code{pheatmap}}
}
\description{
Plot heatmap with enrichment scores by contrast and gene set from \code{\link{gage_all}}
}
\examples{
\dontrun{
  data(msig)
  x <- gage_all(res, msig$KEGG)
  plot_gage(x)
 }
}
\author{
Chris Stubben
}
