% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot heatmap of top genes}
\usage{
plot_genes(x, intgroup, output = "pheatmap", palette = "RdBu",
  dendsort = TRUE, midpoint0 = TRUE, max_scale = NA, border = NA, ...)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{intgroup}{one or more column names for pheatmap \code{annotation_col} bar}

\item{output}{pheatmap or d3heatmap}

\item{palette}{a color palette name or vector of colors, diverging color palettes are reversed}

\item{dendsort}{reorder branches using \code{dendsort} package}

\item{midpoint0}{center color scale midpoint at zero}

\item{max_scale}{min and max value for color scale, default is max(abs(range(x)))}

\item{border}{pheatmap border color, default none}

\item{\dots}{additional options passed to \code{pheatmap} or \code{d3heatmap}}
}
\value{
A pheatmap or d3heatmap
}
\description{
Plot heatmap of top genes with annoation bars, reordered branches, and color scale midpoint at zero.
}
\examples{
data(pasilla)
 x <- top_counts(pasilla$results, pasilla$rlog)
plot_genes(x)
plot_genes(x, "condition", palette="RdBu")
plot_genes(top_counts(pasilla$results, pasilla$rlog, sort_fc=TRUE) )
plot_genes(top_counts(pasilla$results, pasilla$rlog, top=200), output = "d3")
}
\author{
Chris Stubben
}

