% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot counts for many genes}
\usage{
plot_interactions(x, intgroups, ylab = "count", scaled = TRUE,
  n = 40, reorder)
}
\arguments{
\item{x}{a tibble from \code{\link{top_counts}}}

\item{intgroups}{two column names in \code{attr(x, "colData")} to use for grouping}

\item{ylab}{y-axis label}

\item{scaled}{scale counts}

\item{n}{total number of genes to plot}

\item{reorder}{level names to reorder factor levels on x-axis}
}
\value{
A ggplot
}
\description{
Plot counts for many genes
}
\examples{
\dontrun{
 x <- top_counts( res, rld, top=25)
 plot_interactions(x,  c("time", "trt"), ylab = "scaled rlog")
}
}
\author{
Chris Stubben
}
