% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Highchart version of MA-plot}
\usage{
plot_ma(
  res,
  signif = 0.05,
  baseMean = 10000,
  foldchange = 2,
  radius = 3,
  size = 1,
  alpha = 0.3,
  sig_alpha = TRUE,
  ylab = "Log2 Fold Change",
  ggplot = TRUE,
  ...
)
}
\arguments{
\item{res}{Annotated DESeq results table from results_all}

\item{signif}{significance level, default 0.05}

\item{baseMean}{normalized count cutoff for labeling points, default 10000}

\item{foldchange}{absolute value of log2 fold change cutoff for labeling
points, default 2}

\item{radius}{highchart point size, default 3}

\item{size}{ggplot point size, default 1}

\item{alpha}{ggplot alpha transparency, default 0.3}

\item{sig_alpha}{use transparency for significant genes, default TRUE}

\item{ylab}{y-axis label}

\item{ggplot}{plot ggplot version}

\item{\dots}{other options like width passed to \code{hc_chart}}
}
\value{
A highchart or ggplot. Only points above the baseMean or log2 fold
change cutoffs have mouseover labels.
}
\description{
Plot mean normalized counts and fold changes in an interactive MA-plot
}
\examples{
plot_ma(pasilla$results, ggplot = TRUE)
}
\author{
Chris Stubben
}
