% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma.R
\name{plot_ma}
\alias{plot_ma}
\title{Highchart version of MA-plot}
\usage{
plot_ma(res, baseMean = 10000, foldchange = 2, radius = 3,
  ggplot = FALSE, ...)
}
\arguments{
\item{res}{Annotated DESeq results table from results_all}

\item{baseMean}{normalized count cutoff for labeling points, default 10000}

\item{foldchange}{absolute value of log2 fold change cutoff for labeling points, default 2}

\item{radius}{highchart point size, default 3}

\item{ggplot}{plot ggplot version}

\item{\dots}{other options like width passed to \code{hc_chart}}
}
\value{
A highchart or ggplot. Only points above the baseMean or log2 fold change cutoffs have mouseover labels.
}
\description{
Plot mean normalized counts and fold changes in an interactive MA-plot
}
\examples{
\dontrun{
plot_ma(res)
}
}
\author{
Chris Stubben
}
