% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Highchart version of plotPCA in DESeq2}
\usage{
plot_pca(object, intgroup = "trt", tooltip, ntop = 500, relevel,
  pc = c(1, 2), scale = FALSE, ggplot = FALSE, ...)
}
\arguments{
\item{object}{a matrix, ExpressionSet or DESeqTransform object}

\item{intgroup}{a character vector of names in pData(x) or colData(x) for grouping, default 'trt'}

\item{tooltip}{a character vector of names in pData(x) or colData(x) for tooltip display,
default displays the object column names}

\item{ntop}{number of top variable genes to use for principal components}

\item{relevel}{reorder intgroup levels, default is alphabetical}

\item{pc}{a vector of components to plot, default 1st and 2nd}

\item{scale}{option to scale variables in prcomp, default FALSE}

\item{ggplot}{plot ggplot version}

\item{\dots}{additional options passed to \code{hc_chart}}
}
\value{
A highchart
}
\description{
Highchart version of sample PCA plot
}
\examples{
data(pasilla)
plot_pca(pasilla$rlog, "condition", tooltip=c("file", "type"))
plot_pca(pasilla$rlog, c("condition", "type"))
}
\author{
Chris Stubben
}
