% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(res, pvalue_cutoff, foldchange_cutoff, max_pvalue = 200,
  radius = 3, ggplot = TRUE, palette = "RdBu", ...)
}
\arguments{
\item{res}{Annotated DESeq results table from results_all.}

\item{pvalue_cutoff}{label points above this cutoff on a -10 log10 y-axis
scale, the default is 1.3 (padj = 0.05) for highcharts and no labels for
ggplot}

\item{foldchange_cutoff}{absolute value of log2 fold change cutoff for
labeling highchart points, default 2}

\item{max_pvalue}{y-axis limit, maximum value on a -10 log10 y-axis scale,
the default is 200 (padj < 1e-200), so genes below this cutoff are assigned
the maximum p-value.}

\item{radius}{point size, default 3}

\item{ggplot}{plot ggplot version}

\item{palette}{RColorBrewer palette name, vector of colors, or "RdGn" for
ggplot}

\item{\dots}{other options like width passed to \code{hc_chart}}
}
\value{
A highchart or ggplot.
}
\description{
Plot fold changes and adjusted p-values in an interactive volcano plot or
ggplot
}
\details{
For ggplot, the results should not be sorted by p-value or fold change to
avoid stacking close overlapping points.  Labels are added using
\code{ggrepel}, so avoid labeling too many points (200 is the limit).
}
\examples{
\dontrun{
plot_volcano(res, ggplot=TRUE, pvalue=10)
}
}
\author{
Chris Stubben
}
