% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Highchart version of volcano plot}
\usage{
plot_volcano(res, padj = 0.05, log2FoldChange = 2, max_pvalue = 100,
  radius = 4, ggplot = FALSE, ...)
}
\arguments{
\item{res}{Annotated DESeq results table from results_all}

\item{padj}{p-value cutoff for labeling points, default 0.05}

\item{log2FoldChange}{absolute value of log2 fold change cutoff for labeling points, default 2}

\item{max_pvalue}{y-axis range from 0 to max_pvalue, default 100 (padj = 1e-100)}

\item{radius}{point size, default 4}

\item{ggplot}{plot ggplot version}

\item{\dots}{other options like width passed to \code{hc_chart}}
}
\value{
A highchart. Only points above the adjusted p-value and log2 fold change cutoffs have mouseover labels.
}
\description{
Plot fold changes and adjusted p-values in an interactive volcano plot
}
\examples{
\dontrun{
plot_volcano(res)
}
}
\author{
Chris Stubben
}
