% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomart.R
\name{read_biomart}
\alias{read_biomart}
\title{Read biomart annotations}
\usage{
read_biomart(
  dataset = "human",
  attributes,
  host = "www.ensembl.org",
  version = NULL,
  patch = FALSE,
  list = NULL,
  ...
)
}
\arguments{
\item{dataset}{The first letter of genus and full species name like btaurus,
ecaballus, sscrofa from \code{listEnsembl}. A few common names are accepted
for human, mouse, rat, fruitfly, yeast and zebrafish.}

\item{attributes}{vector of column names to pass to \code{getBM}, default
ensembl_gene_id, external_gene_name, gene_biotype, chromosome_name,
start_position, end_position, strand, description and transcript_count}

\item{host}{host for connection, default www.ensembl.org}

\item{version}{Ensembl version for previous releases}

\item{patch}{Keep features on patches starting with CHR_, default FALSE}

\item{list}{return a list of either datasets, attributes or filters only.}

\item{\dots}{additional options passed to \code{getBM} or
\code{listAttributes}}
}
\value{
A tibble
}
\description{
Read Ensembl annotations using \code{biomaRt} library
}
\note{
Many attributes like entrezgene have a 0 to many relationship with
ensembl_gene_id causing duplicate ensembl ids to be added.
}
\examples{
\dontrun{
x1 <- read_biomart( list= "datasets")
x1
# Download latest version
 mmu <- read_biomart("mouse")
 # Download mouse genes and human homologs
 x1 <- read_biomart("mouse", list= "attributes")
 dplyr::count(x1, page)
 filter(x1, grepl("hsapiens_homolog", name))
 mmu_homologs <- read_biomart("mouse",  attributes = c("ensembl_gene_id",
   "hsapiens_homolog_ensembl_gene", "hsapiens_homolog_associated_gene_name",
   "hsapiens_homolog_perc_id"))
 # Human genes with SignalP
 x2 <- read_biomart("human", list= "filters")
 filter(x2, grepl("signal", name))
 hsa_signalp <- read_biomart("human", attributes = c("ensembl_transcript_id",
  "ensembl_gene_id",  "external_gene_name", "signalp_start",  "signalp_end"),
  filter="with_signalp", values=TRUE)
}
}
\author{
Chris Stubben
}
