% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_all.R
\name{results_all}
\alias{results_all}
\title{Extract and annotate all results from a DESeq analysis}
\usage{
results_all(object, biomart, vs = "all", vs2 = TRUE, alpha = 0.05,
  add_columns, other_columns, simplify = TRUE, ...)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{biomart}{annotations from \code{read_biomart} with column 1 matching row names in results}

\item{vs}{either compare all vs. all (default) or a specific treatment vs. all, or see note.}

\item{vs2}{position of specific treatment in contrast vector, set FALSE for specific treatment vs all}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{add_columns}{a vector of biomart columns to add to result table, default
gene_name, biotype, chromosome, start and description}

\item{other_columns}{a vector of additional columns in biomart table}

\item{simplify}{return a tibble if only 1 contrast present}

\item{\dots}{additional options passed to \code{results}}
}
\value{
A list of tibbles for each contrast
}
\description{
Extract all possible contrasts and annotate result tables from a DESeq object.
Currently supports simple designs with a single variable.
}
\note{
If you combine factors of interest into a single group following section 3.3 in the DESeq2 vignette,
 you can set vs = "combined1", "combined2", or "combined" to limit the comparisons.
 For example, if you combine 3 cell types and 2 time points, then the default
returns 18 contrasts,  "combined1" returns 3 contrasts comparing time within cell types, combined2 returns 6 contrasts
comparing cell types at the same time point or "combined" to return both (9 total).
Factors should be separated by spaces in the combined treatment group for parsing.
}
\examples{
\dontrun{
data(pasilla)
data(fly)
res <- results_all(pasilla$dds, fly)
res
# Set factor levels in the DESeq object to change contrast order
#  since results_all uses combn on levels
dds$trt <- factor(dds$trt, levels=c("heart", "lung", "control"))
apply( combn(c("heart", "lung", "control"), 2), 2, paste, collapse= " vs. ")
}
}
\author{
Chris Stubben
}

