% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_all.R
\name{results_all}
\alias{results_all}
\title{Extract all results from a DESeq analysis}
\usage{
results_all(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{\dots}{additional options passed to \code{results}}
}
\value{
A list of DESeqResults objects for each contrast
}
\description{
Extract all possible contracts and result tables from a DESeq analysis
Currently supports simple designs with a single variable.
}
\examples{
\dontrun{
  res <- results_all(dds)
}
}
\author{
Chris Stubben
}

