% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_all.R
\name{results_all}
\alias{results_all}
\title{Extract and annotate all results from a DESeq analysis}
\usage{
results_all(object, biomart, vs = "all", subset, relevel, alpha = 0.05,
  add_columns, trt, lfcShrink = TRUE, simplify = TRUE, ...)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{biomart}{annotations from \code{read_biomart} with column 1 matching row names in results}

\item{vs}{either compare all vs. all (default) or all vs specific treatment, or see note.}

\item{subset}{index to subset all pairwise comparisons, try \code{combn(sort(samples$trt),2)}}

\item{relevel}{Levels to compare, if missing then levels(dds$trt)}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{add_columns}{a vector of biomart columns to add to result table, default
gene_name, biotype, chromosome, description and human_homolog if present}

\item{trt}{Compare groups within trt group, default is first term in the design formula}

\item{lfcShrink}{shrink fold changes using \code{lfcShrink} for DESeq2 version >= 1.16}

\item{simplify}{return a tibble if only 1 contrast present}

\item{\dots}{additional options passed to \code{results}}
}
\value{
A list of tibbles for each contrast
}
\description{
Extract all possible contrasts and annotate result tables from a DESeq object.
Currently supports simple designs with a single variable.
}
\details{
Uses the first treatment in the design formula for contrasts.  Levels are
taken from levels(dds$trt) or use the levels option to re-order or compare a subset of levels.
}
\note{
If you combine factors of interest into a single group following section 3.3 in the DESeq2 vignette,
 you can set vs = "combined" to limit the comparisons.  See \code{\link{check_contrasts}} for details.
 If you combine 3 cell types and 2 treatments, then the default
returns 18 contrasts while "combined" returns 3 contrasts comparing treatment within cell types (first group).
Factors should be separated by space, dash or underscore in the combined treatment group for parsing.
}
\examples{
\dontrun{
data(pasilla)
data(fly)
res <- results_all(pasilla$dds, fly)
res
# Relevel to change default contrast order in dds$trt
levels(dds$trt)
results_all(dds, mmu, relevel=c("heart", "lung", "control"))
}
}
\author{
Chris Stubben
}
