% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_all.R
\name{results_all}
\alias{results_all}
\title{Extract and annotate all results from a DESeq analysis}
\usage{
results_all(
  object,
  biomart,
  vs = "all",
  subset,
  relevel,
  alpha = 0.05,
  add_columns,
  trt,
  lfcShrink = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{a DESeqDataSet}

\item{biomart}{annotations from \code{read_biomart} with column 1 matching row names in results}

\item{vs}{either compare all vs. all (default) or all vs specific treatment like control, or see note.}

\item{subset}{index to subset all pairwise comparisons, try \code{combn(sort(samples$trt),2)}}

\item{relevel}{Levels to compare, if missing then levels(dds$trt)}

\item{alpha}{the significance cutoff for the adjusted p-value cutoff (FDR)}

\item{add_columns}{a vector of biomart columns to add to result table, default
gene_name, biotype, chromosome, description and human_homolog if present}

\item{trt}{Compare groups within trt group, default is first term in the design formula}

\item{lfcShrink}{shrink fold changes using \code{lfcShrink} with type = "normal"}

\item{simplify}{return a tibble if only 1 contrast present}

\item{\dots}{additional options passed to \code{results}}
}
\value{
A list of tibbles for each contrast
}
\description{
Extract all possible contrasts and annotate result tables from a DESeq object.
Currently supports simple designs with a single variable.  Following
DESeq2 conventions, control group should be listed first and then pairwise
comparisons are selected in reverse order, so A, B, C levels will return
C vs. B, C vs. A, and B vs. A
}
\note{
If you combine factors of interest into a single group following section 3.3 in the DESeq2 vignette,
 you can set vs = "combined" to limit the comparisons.  See \code{\link{check_contrasts}} for details.
}
\examples{
\dontrun{
library(hciRdata)
res <- results_all(pasilla$dds, fly98)
res
}
}
\author{
Chris Stubben
}
