% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_counts.R
\name{sort_counts}
\alias{sort_counts}
\title{Sort columns of a count matrix by rows in sample data}
\usage{
sort_counts(count_tbl, sample_tbl, id = 1)
}
\arguments{
\item{count_tbl}{count matrix, data.frame or tibble}

\item{sample_tbl}{sample tabl with a column containing the columns names in count matrix}

\item{id}{position of column in sample data that matches column names in the
count table, default is the first column}
}
\value{
A count matrix with columns re-ordered to match the sample table
}
\description{
\code{DESeqDataSetFromMatrix} requires count columns sorted by the rows in sample data and only checks
if sample row names match count column names (and tibbles do not have row names).  This matches column
names in a count table to the first column in sample data and reorders columns to match.
}
\examples{
\dontrun{
count_tbl <- sort_counts(counts, samples)
}
}
\author{
Chris Stubben
}
