% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_samples.R
\name{sort_samples}
\alias{sort_samples}
\title{Sort sample names}
\usage{
sort_samples(samples)
}
\arguments{
\item{samples}{a vector of sample names}
}
\value{
A vector
}
\description{
Sort character vectors by removing non-numeric characters.  This works for very simple names,
 try \code{mixedsort} in \code{ggtools} for more complicated examples
}
\examples{
 samples <- c("135X1", "135X2", "135X11")
 sort(samples)
 sort_samples(samples)
}
\author{
Chris Stubben
}

