% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_deseq.R
\name{summary_deseq}
\alias{summary_deseq}
\title{Summarize DESeq2 results}
\usage{
summary_deseq(object)
}
\arguments{
\item{object}{a DESeqResults object}
}
\value{
A data.frame
}
\description{
Summarize DESeq2 results into a \code{data.frame}
}
\examples{
summary_deseq(pasilla$results)
}
\author{
Chris Stubben
}
