% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_counts.R
\name{top_counts}
\alias{top_counts}
\title{Get a count matrix of top genes}
\usage{
top_counts(res, rld, by = "id", filter = TRUE, col_names,
  row_names = "gene_name", collapse, ...)
}
\arguments{
\item{res}{an annotated DESeq2 results file}

\item{rld}{rlog or other counts in DESeqTransform object}

\item{by}{join count rownames by this column number in results, default id}

\item{filter}{filter the results using \code{\link{top_genes}}}

\item{col_names}{optionally replace sample IDs with another column in colData(rld)}

\item{row_names}{a column name in results to use as row names, default gene_name}

\item{collapse}{collapse replicates and group by column means in colData(rld)}

\item{\dots}{additional options passed to \code{\link{top_genes}}}
}
\value{
A tibble with colData attribute
}
\description{
Join DESeq results to rlog or other counts, mainly for plotting gene heatmaps.
}
\examples{
\dontrun{
 x <- top_counts(res[[1]], rld)
 pheatmap(as_matrix(x))
 plot_genes(x, "trt")
}
}
\author{
Chris Stubben
}
