% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_counts.R
\name{top_counts}
\alias{top_counts}
\title{Get a count matrix of top genes}
\usage{
top_counts(res, rld, top = 40, alpha = 0.05, basemean, log2FC, col_names,
  trt, row_names = "gene_name", difference = TRUE, sort_fc = FALSE, ...)
}
\arguments{
\item{res}{an annotated DESeq2 results file}

\item{rld}{rlog or other counts in DESeqTransform object}

\item{top}{Number of top genes to display in matrix}

\item{basemean}{basemean cutoff}

\item{log2FC}{absolute value log2 fold change cutoff}

\item{col_names}{a column name in colData(rld) to use as column names}

\item{row_names}{a column name in results to use as row names, default gene_name}

\item{difference}{Subtract the row mean from counts to display the difference from the gene's average}

\item{sort_fc}{Sort top genes by fold changes and get top n/2 up and down-regulated, default is to sort by adjusted p-value}

\item{by}{join count rownames by this column number in results, default 1}
}
\value{
A tibble with colData attribute
}
\description{
Join top genes in DESeq results to rlog or other counts, mainly for plotting gene heatmaps.
}
\examples{
\dontrun{
 x <- top_counts(res, rld)
 pheatamp(x)
 pheatmap_genes(x)
 # or with annotation bar, re-ordered branches and color scale midpoint at zero
 pheatmap_genes(x, "Trt", border=NA )
}
}
\author{
Chris Stubben
}

