% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_deseq.R
\name{write_deseq}
\alias{write_deseq}
\title{Write DESeq results}
\usage{
write_deseq(
  result_all,
  dds,
  rld,
  biomart,
  sets,
  fpkm,
  text_files = FALSE,
  cutoff,
  file = "DESeq.xlsx",
  ...
)
}
\arguments{
\item{result_all}{a list from \code{results_all}}

\item{dds}{a DESeqDataset object with count tables}

\item{rld}{a DESeqTransform obect with rlog values}

\item{biomart}{annotations from \code{read_biomart}}

\item{sets}{table with set intersections, optional}

\item{fpkm}{matrix of fpkm values, optional}

\item{text_files}{write results to separate txt files, mainly for IPA input}

\item{cutoff}{if text_files is TRUE, only write genes below this p-value cutoff}

\item{file}{file name}

\item{\dots}{additional options passed to \code{annotate_results}}
}
\value{
A tibble
}
\description{
Write DESeq result files, raw counts, rlog values, normalized counts to an Excel file.
}
\examples{
\dontrun{
 write_deseq(res, dds, rld, human98)
}
}
\author{
Chris Stubben
}
